requirements.txt
cocos2d==0.6.5
pkg-resources==0.0.0
pyglet==1.2.4
six==1.11.0
